Apolo/requirements.txt
asgiref==3.2.3
astroid==2.3.3
attrs==19.3.0
autopep8==1.5
auxlib==0.0.43
backcall==0.1.0
bleach==3.1.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
decorator==4.4.1
defusedxml==0.6.0
Django==1.11.20
django-cors-headers==3.2.1
djangorestframework==3.8.2
entrypoints==0.3
google-api-python-client==1.7.11
google-auth==1.10.0
google-auth-httplib2==0.0.3
httplib2==0.15.0
idna==2.8
importlib-metadata==1.4.0
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.15.2
Jinja2==2.10.3
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.1.0
nbconvert==5.6.1
nbformat==5.0.3
notebook==6.0.2
numpy==1.18.1
pandocfilters==1.4.2
parso==0.5.2
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==2.0.10
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
Pygments==2.5.2
pylint==2.4.4
pylint-django==2.0.13
pylint-plugin-utils==0.6
pyrsistent==0.15.7
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
pywin32==227
pywinpty==0.5.7
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.22.0
rsa==4.0
scikit-learn==0.22.1
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
sqlparse==0.3.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==1.0.0
